+++ includes/admin-page.php
@@ -66,12 +66,6 @@
                     <th><?php _e( 'Key Prefix:', 'redis-cache' ); ?></th>
                     <td>
                         <code><?php echo esc_html( $redisPrefix ); ?></code>
-
-                        <?php if ( strlen( (string) $redisPrefix ) > 20 || ! preg_match( '/[\w-]/', $redisPrefix ) ) : ?>
-                            <p class="description" style="color: #d54e21;">
-                                <?php _e( 'Consider using a shorter, human-readable prefix.', 'redis-cache' ); ?>
-                            </p>
-                        <?php endif; ?>
                     </td>
                 </tr>
             <?php endif; ?>